fix gem issues
