+++ .travis.yml
@@ -1,9 +1,9 @@
 language: ruby
-before_install:
-  - travis_retry gem install bundler
 rvm:
+  - 1.9.3
   - 2.0.0
-  - 2.1
+  - 2.1.5
+  - 2.2.0
   - ruby-head
   - jruby
   - rbx-2
