+++ Gemfile
@@ -1,6 +1,4 @@
-source 'http://rubygems.org'
-
-gem 'rdoc'
+source 'https://rubygems.org'
 
 # Specify your gem's dependencies in guard-konacha-rails.gemspec
 gemspec
