+++ Readme.md
@@ -1,4 +1,4 @@
-# Rediska
+# guard-konacha-rails
 
 [![Gem Version](https://badge.fury.io/rb/guard-konacha-rails.png)](http://badge.fury.io/rb/guard-konacha-rails)
 [![Build Status](https://travis-ci.org/lbeder/guard-konacha-rails.png)](https://travis-ci.org/lbeder/guard-konacha-rails)
@@ -12,40 +12,19 @@ Automatically run your [Konacha](https://github.com/jfirebaugh/konacha) tests th
 If you are using bundler add guard-konacha-rails to your Gemfile:
 
 ``` ruby
-gem 'guard-konacha-rails'
+group :development do
+  gem 'guard-konacha-rails'
+end
 ```
 
-Then run:
+Then add the configuration to the Guardfile:
 
 ```bash
-$ bundle install
+$ bundle exec guard init konacha-rails
 ```
 
-Otherwise install the gem:
+Then to run:
 
 ```bash
-$ gem install guard-konacha-rails
+$ bundle exec guard
 ```
-
-## License
-
-The MIT License (MIT)
-
-Copyright (c) 2013
-
-Permission is hereby granted, free of charge, to any person obtaining a copy of
-this software and associated documentation files (the "Software"), to deal in
-the Software without restriction, including without limitation the rights to
-use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
-the Software, and to permit persons to whom the Software is furnished to do so,
-subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
-FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
-CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
