+++ guard-konacha-rails.gemspec
@@ -1,9 +1,11 @@
-$:.push File.expand_path('../lib', __FILE__)
-require 'guard/konacha/rails/version'
+# coding: utf-8
+lib = File.expand_path('../lib', __FILE__)
+$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
+require 'guard/konacha-rails/version'
 
 Gem::Specification.new do |s|
   s.name        = 'guard-konacha-rails'
-  s.version     = Guard::KonachaRails::VERSION
+  s.version     = Guard::KonachaRailsVersion::VERSION
   s.platform    = Gem::Platform::RUBY
   s.authors     = ['Leonid Beder']
   s.email       = ['leonid.beder@gmail.com']
@@ -12,13 +14,13 @@ Gem::Specification.new do |s|
   s.summary     = 'Guard plugin for the konacha testing framework'
   s.description = 'Guard plugin for the konacha testing framework.'
 
-  s.files         = `git ls-files`.split("\n")
-  s.test_files    = `git ls-files -- {test,spec,features}/*`.split("\n")
-  s.executables   = `git ls-files -- bin/*`.split("\n").map{ |f| File.basename(f) }
-  s.require_paths = ['lib']
+  s.files         = `git ls-files -z`.split("\x0")
+  s.executables   = s.files.grep(%r{^bin/}) { |f| File.basename(f) }
+  s.test_files    = s.files.grep(%r{^(test|spec|features)/})
+  s.require_paths = ["lib"]
 
   s.add_dependency 'guard', '>= 2'
-  s.add_dependency 'rails', '>= 4.2'
+  s.add_dependency 'rails', '>= 4.0'
   s.add_dependency 'konacha', '>= 3.0'
   s.add_dependency 'guard-compat'
