+++ lib/guard/konacha-rails/templates/Guardfile
@@ -1,10 +1,10 @@
-### Guard::Konacha
+### Guard::KonachaRails
 #  available options:
 #  - :run_all_on_start, defaults to true
 #  - :notification, defaults to true
 #  - :rails_environment_file, location of rails environment file,
 #    should be able to find it automatically
-guard :konacha do
-  watch(%r{^app/assets/javascripts/(.*)\.js(\.coffee)?$}) { |m| "#{m[1]}_spec.js" }
-  watch(%r{^spec/javascripts/.+_spec(\.js|\.js\.coffee)$})
+guard "konacha-rails" do
+  watch(%r{^app/assets/javascripts/(.*)\.(js|coffee)?$}) { |m| "#{m[1]}_spec.js" }
+  watch(%r{^(test|spec)/javascripts/.+_(test|spec)\.(js|coffee)$})
 end
