+++ spec/guard/konacha-rails/rails/formatter_spec.rb
