+++ spec/spec_helper.rb
@@ -3,7 +3,7 @@ ENV['RAILS_ENV'] ||= 'test'
 require 'rspec'
 require 'coveralls'
 require 'timecop'
-require 'guard/konacha'
+require 'guard/konacha-rails'
 require 'guard/compat/test/helper'
 
 Coveralls.wear!
